move src/utilities/error_utils.h to include/utilities/error.hpp following cudf (and users face exception classes, so this header should be part of public API)
