+++ cpp/src/community/ECG.cu
@@ -18,7 +18,7 @@
 
 #include <rmm/thrust_rmm_allocator.h>
 #include <thrust/random.h>
-#include <utilities/error_utils.h>
+#include <utilities/error.hpp>
 #include <converters/permute_graph.cuh>
 #include <ctime>
 #include "utilities/graph_utils.cuh"
