+++ cpp/src/community/louvain.cpp
@@ -23,7 +23,7 @@
 
 #include <community/louvain_kernels.hpp>
 
-#include "utilities/error_utils.h"
+#include "utilities/error.hpp"
 
 namespace cugraph {
