+++ cpp/src/components/connectivity.cu
@@ -8,7 +8,7 @@
 #include <graph.hpp>
 #include <iostream>
 #include <type_traits>
-#include "utilities/error_utils.h"
+#include "utilities/error.hpp"
 #include "utilities/graph_utils.cuh"
 
 #include "topology/topology.cuh"
