+++ cpp/src/db/db_object.cu
@@ -16,7 +16,7 @@
 
 #include <rmm/thrust_rmm_allocator.h>
 #include <thrust/binary_search.h>
-#include <utilities/error_utils.h>
+#include <utilities/error.hpp>
 #include <cub/device/device_run_length_encode.cuh>
 #include <db/db_object.cuh>
 #include <sstream>
