+++ cpp/src/layout/exact_fa2.hpp
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <rmm/thrust_rmm_allocator.h>
-#include <utilities/error_utils.h>
+#include <utilities/error.hpp>
 
 #include <stdio.h>
 #include <converters/COOtoCSR.cuh>
