+++ cpp/src/link_analysis/pagerank.cu
@@ -23,7 +23,7 @@
 #include "cub/cub.cuh"
 
 #include <rmm/thrust_rmm_allocator.h>
-#include <utilities/error_utils.h>
+#include <utilities/error.hpp>
 
 #include <graph.hpp>
 #include "utilities/graph_utils.cuh"
