+++ cpp/src/link_prediction/jaccard.cu
@@ -20,7 +20,7 @@
  * ---------------------------------------------------------------------------**/
 
 #include <rmm/thrust_rmm_allocator.h>
-#include <utilities/error_utils.h>
+#include <utilities/error.hpp>
 #include "graph.hpp"
 #include "utilities/graph_utils.cuh"
