+++ cpp/src/sort/bitonic.cuh
@@ -34,7 +34,7 @@
 #include <thrust/scan.h>
 
 #include <rmm/thrust_rmm_allocator.h>
-#include <utilities/error_utils.h>
+#include <utilities/error.hpp>
 
 namespace cugraph {
 namespace sort {
