+++ cpp/src/traversal/traversal_common.cuh
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <cub/cub.cuh>
-#include "utilities/error_utils.h"
+#include "utilities/error.hpp"
 
 #define MAXBLOCKS 65535
 #define WARP_SIZE 32
