+++ cpp/src/traversal/two_hop_neighbors.cu
@@ -20,7 +20,7 @@
  * ---------------------------------------------------------------------------**/
 
 #include <rmm/thrust_rmm_allocator.h>
-#include <utilities/error_utils.h>
+#include <utilities/error.hpp>
 #include <algorithms.hpp>
 #include <graph.hpp>
 #include "two_hop_neighbors.cuh"
