+++ cpp/src/utilities/graph_utils.cuh
@@ -24,7 +24,7 @@
 #include <thrust/transform.h>
 
 #include <rmm/thrust_rmm_allocator.h>
-#include <utilities/error_utils.h>
+#include <utilities/error.hpp>
 
 namespace cugraph {
 namespace detail {
