+++ cpp/tests/traversal/bfs_test.cu
@@ -21,7 +21,6 @@
 #include <algorithms.hpp>
 
 #include <rmm/thrust_rmm_allocator.h>
-#include <utilities/error_utils.h>
 
 #include "gtest/gtest.h"
